requirements.txt
tensorflow==2.15.0
matplotlib==3.8.2
fastapi===0.104.1
pydantic
uvicorn
python-multipart
python-dotenv